+++ src/haystack/split/share.cr
@@ -2,5 +2,5 @@ struct Haystack::Split::Share
   include Hapi::Resource
 
   getter subaccount : Subaccount
-  getter share : Int64
+  getter share : Int32
 end
